Refactor bot into modular scripts
